🧪 Add unit tests for svdtriplet function in algebra.py
